# RavenDB-23077 - PerCoreContainer performance improvements