### Issue link

https://issues.hibernatingrhinos.com/issue/RavenDB-23077/PerCoreContainer-performance-improvements

### Additional description

Each time an item is pushed or pulled from the `PerCoreContainer`, the entire core array is scanned (worst case) to find a null slot or a non-null item. This introduces unnecessary looping, especially if the array is mostly empty or full.

Now, we check if the array is full before attempting `TryPush` and if it's empty before attempting `TryPull`.

The change significantly improves performance in both single-threaded and multi-threaded scenarios:

Single-threaded push:
Old: `2608ms`
New: `317ms`
**~8x faster**

Multi-threaded push/pull:
Old: `10304ms`
New: `3118ms`
**~3.2x faster**

### Type of change

- [ ] Bug fix
- [ ] Regression bug fix
- [x] Optimization
- [ ] New feature

### How risky is the change?

- [x] Low 
- [ ] Moderate 
- [ ] High
- [ ] Not relevant

### Backward compatibility

- [x] Non breaking change
- [ ] Ensured. Please explain how has it been implemented?
- [ ] Breaking change
- [ ] Not relevant

### Is it platform specific issue?

- [ ] Yes. Please list the affected platforms.
- [x] No

### Documentation update

- [ ] This change requires a documentation update. Please mark the issue on YouTrack using `Documentation Required` tag.
- [x] No documentation update is needed 

### Testing by Contributor

- [ ] Tests have been added that prove the fix is effective or that the feature works
- [ ] Internal classes added to the test class (e.g. entity or index definition classes) have the lowest possible access modifier (preferable `private`) 
- [x] It has been verified by manual testing

### Testing by RavenDB QA team

- [ ] This change requires a special QA testing due to possible performance or resources usage implications (CPU, memory, IO). Please mark the issue on YouTrack using `QA Required` tag.
- [x] No special testing by RavenDB QA team is needed

### Is there any existing behavior change of other features due to this change?

- [ ] Yes. Please list the affected features/subsystems and provide appropriate explanation
- [x] No

### UI work

- [ ] It requires further work in the Studio. Please mark the issue on YouTrack using `Studio Required` tag.
- [x] No UI work is needed